+++ src/com/zimplistic/rotimatic/testcases/webDriverResponse.java
@@ -1,3 +1,4 @@
+package com.zimplistic.rotimatic.testcases;
 
 public class webDriverResponse {
